fix(contracts): generator umów — stawka progresywna, adres do korespondencji, grosze w stawce ramowej
Trzy zgłoszenia z modułu Kontrakty:

1. Generator umów B2B, „Warunki umowy": stawka progresywna — kilka etapów
   (kwota + „Obowiązuje od" / „Obowiązuje do") w ramach jednej umowy.
   Zdanie o stawce (§5) w szablonach DOCX/HTML buduje `b2b.rate_clause`
   (rate_clause.py): pojedyncza stawka renderuje się identycznie jak
   dotychczas, etapy → wyliczenie okresów, każda kwota z własnym „słownie".
   Etapy z /generate trafiają też do harmonogramu candidate_rate_schedule
   (spójnie z „Dodaj stawkę progresywną" w edycji kontraktu).

2. Generator umów B2B, „Dane Partnera": wypełniony „Adres do korespondencji"
   pojawia się w komparycji wygenerowanej umowy (PL/EN, DOCX + podgląd HTML)
   po REGON. Fragment warunkowy — bez adresu komparycja bez zmian;
   whitespace-only wartość jest normalizowana do braku adresu.

3. Edycja kontraktu: „Stawka z umowy ramowej" i „Widełki docelowe" przyjmują
   grosze (np. 215,60) — kolumny INTEGER → NUMERIC(12,2) (migracja 0157 +
   idempotentny safety-net w entrypoint.sh), schematy int → float. Koniec
   „Request failed with status code 422".

+++ backend/app/models/contract.py
@@ -128,7 +128,9 @@ class Contract(Base, TimestampMixin):
     rate_client: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 3))
     # Stawka z umowy ramowej (MSA) — wartość referencyjna uzgodniona w umowie
     # ramowej z klientem. NIE wchodzi do liczenia marży (to baseline/ceiling).
-    framework_rate: Mapped[Optional[int]] = mapped_column(Integer)
+    # Numeric(12,2): stawki ramowe bywają z groszami (np. 215,60) — Integer
+    # odrzucał je 422-ką na schemacie (migracja 0157).
+    framework_rate: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2))
     currency: Mapped[str] = mapped_column(String(3), default="PLN")
 
     # Jednostka stawki (godz. / dzień / mies.) + liczba godzin billingowych (dla stawki godzinowej).
@@ -191,9 +193,14 @@ class Contract(Base, TimestampMixin):
     terminated_at: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
 
     # Desired rate range we want to achieve on this contract (used by benchmark
-    # comparison and by sales during renegotiation).
-    target_rate_min: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    target_rate_max: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    # comparison and by sales during renegotiation). Numeric(12,2) — grosze jak
+    # w framework_rate (migracja 0157).
+    target_rate_min: Mapped[Optional[Decimal]] = mapped_column(
+        Numeric(12, 2), nullable=True
+    )
+    target_rate_max: Mapped[Optional[Decimal]] = mapped_column(
+        Numeric(12, 2), nullable=True
+    )
 
     # End of the client's purchase order — often earlier than our contract with
     # the consultant. Drives proactive reminders so we can react before the
